Migrate Freshservice MCP server to static metadata pattern
Move from inline server definition to the new metadata pattern:
- Create metadata.ts with createToolsRecord for type-safe tool definitions
- Create tools/index.ts with ToolHandlers for exhaustive handler checking
- Extract API client to client.ts
- Move types to types.ts
- Use registerTool pattern in index.ts

This follows the methodology in runbooks/NEW_MCP_SERVER.md
